requirements.txt
Django==4.1

djangorestframework==3.12.0
djoser==2.1.0
djangorestframework_simplejwt==4.8.0
drf-yasg==1.21.3
django-cors-headers==3.13.0